give all items on account creation
