+++ package.json
@@ -38,7 +38,6 @@
   "dependencies": {
     "bcryptjs": "^2.4.3",
     "denque": "^2.1.0",
-    "fs": "^0.0.1-security",
     "hostile": "^1.3.3",
     "md5": "^2.3.0",
     "promise-events": "^0.2.4",
